Revert "status: extract statuscollector"
This reverts commit 27410e796ed3d86406a9979e6aec9919aa0d8bb8.
