+++ daemon/cmd/daemon.go
@@ -17,11 +17,15 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/vishvananda/netlink"
 
+	"github.com/cilium/cilium/api/v1/models"
 	"github.com/cilium/cilium/daemon/cmd/cni"
 	agentK8s "github.com/cilium/cilium/daemon/k8s"
+	"github.com/cilium/cilium/pkg/auth"
 	"github.com/cilium/cilium/pkg/clustermesh"
+	cmtypes "github.com/cilium/cilium/pkg/clustermesh/types"
 	"github.com/cilium/cilium/pkg/controller"
 	linuxdatapath "github.com/cilium/cilium/pkg/datapath/linux"
+	"github.com/cilium/cilium/pkg/datapath/linux/bigtcp"
 	"github.com/cilium/cilium/pkg/datapath/linux/probes"
 	linuxrouting "github.com/cilium/cilium/pkg/datapath/linux/routing"
 	"github.com/cilium/cilium/pkg/datapath/linux/safenetlink"
@@ -35,6 +39,7 @@ import (
 	"github.com/cilium/cilium/pkg/endpoint/regeneration"
 	"github.com/cilium/cilium/pkg/endpointmanager"
 	"github.com/cilium/cilium/pkg/health"
+	hubblecell "github.com/cilium/cilium/pkg/hubble/cell"
 	"github.com/cilium/cilium/pkg/identity"
 	identitycell "github.com/cilium/cilium/pkg/identity/cache/cell"
 	"github.com/cilium/cilium/pkg/identity/identitymanager"
@@ -48,10 +53,12 @@ import (
 	"github.com/cilium/cilium/pkg/kvstore"
 	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/loadbalancer/experimental"
+	"github.com/cilium/cilium/pkg/lock"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/maglev"
 	"github.com/cilium/cilium/pkg/maps/ctmap"
 	"github.com/cilium/cilium/pkg/maps/lbmap"
+	"github.com/cilium/cilium/pkg/maps/policymap"
 	"github.com/cilium/cilium/pkg/metrics"
 	monitoragent "github.com/cilium/cilium/pkg/monitor/agent"
 	"github.com/cilium/cilium/pkg/mtu"
@@ -61,6 +68,7 @@ import (
 	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/policy"
 	policyAPI "github.com/cilium/cilium/pkg/policy/api"
+	"github.com/cilium/cilium/pkg/proxy"
 	"github.com/cilium/cilium/pkg/rate"
 	"github.com/cilium/cilium/pkg/redirectpolicy"
 	"github.com/cilium/cilium/pkg/resiliency"
@@ -82,9 +90,14 @@ type Daemon struct {
 	logger    *slog.Logger
 	clientset k8sClient.Clientset
 	db        *statedb.DB
+	l7Proxy   *proxy.Proxy
 	svc       service.ServiceManager
 	policy    policy.PolicyRepository
 	idmgr     identitymanager.IDManager
+
+	statusCollectMutex lock.RWMutex
+	statusResponse     models.StatusResponse
+	statusCollector    *status.Collector
 
 	monitorAgent monitoragent.Agent
 
@@ -93,6 +106,7 @@ type Daemon struct {
 	devices          statedb.Table[*datapathTables.Device]
 	nodeAddrs        statedb.Table[datapathTables.NodeAddress]
 
+	clusterInfo cmtypes.ClusterInfo
 	clustermesh *clustermesh.ClusterMesh
 
 	mtuConfig mtu.MTU
@@ -105,6 +119,8 @@ type Daemon struct {
 
 	// ipam is the IP address manager of the agent
 	ipam *ipam.IPAM
+
+	policyMapFactory policymap.Factory
 
 	endpointCreator endpointcreator.EndpointCreator
 	endpointManager endpointmanager.EndpointManager
@@ -141,23 +157,30 @@ type Daemon struct {
 	controllers *controller.Manager
 	jobGroup    job.Group
 
+	// BIG-TCP config values
+	bigTCPConfig *bigtcp.Configuration
+
 	// just used to tie together some status reporting
 	cniConfigManager cni.CNIConfigManager
+
+	// authManager for reporting the status of the auth system certificate provider
+	authManager *auth.AuthManager
 
 	// read-only map of all the hive settings
 	settings cellSettings
 
-	bwManager datapath.BandwidthManager
+	// Tunnel-related configuration
+	tunnelConfig tunnel.Config
+	bwManager    datapath.BandwidthManager
 
 	wireguardAgent *wireguard.Agent
 	orchestrator   datapath.Orchestrator
+	hubble         hubblecell.HubbleIntegration
 
 	lrpManager   *redirectpolicy.Manager
 	maglevConfig maglev.Config
 
 	explbConfig experimental.Config
-
-	statusCollector status.StatusCollector
 }
 
 func (d *Daemon) init() error {
@@ -345,11 +368,17 @@ func newDaemon(ctx context.Context, cleaner *daemonCleanup, params *daemonParams
 		policy:            params.Policy,
 		idmgr:             params.IdentityManager,
 		cniConfigManager:  params.CNIConfigManager,
+		clusterInfo:       params.ClusterInfo,
 		clustermesh:       params.ClusterMesh,
 		monitorAgent:      params.MonitorAgent,
 		svc:               params.ServiceManager,
+		l7Proxy:           params.L7Proxy,
+		authManager:       params.AuthManager,
 		settings:          params.Settings,
+		bigTCPConfig:      params.BigTCPConfig,
+		tunnelConfig:      params.TunnelConfig,
 		bwManager:         params.BandwidthManager,
+		policyMapFactory:  params.PolicyMapFactory,
 		endpointCreator:   params.EndpointCreator,
 		endpointManager:   params.EndpointManager,
 		k8sWatcher:        params.K8sWatcher,
@@ -357,11 +386,11 @@ func newDaemon(ctx context.Context, cleaner *daemonCleanup, params *daemonParams
 		ipam:              params.IPAM,
 		wireguardAgent:    params.WGAgent,
 		orchestrator:      params.Orchestrator,
+		hubble:            params.Hubble,
 		lrpManager:        params.LRPManager,
 		maglevConfig:      params.MaglevConfig,
 		explbConfig:       params.ExpLBConfig,
 		ciliumHealth:      params.CiliumHealth,
-		statusCollector:   params.StatusCollector,
 	}
 
 	// initialize endpointRestoreComplete channel as soon as possible so that subsystems
